+++ CHANGELOG.rst
@@ -2,12 +2,19 @@
 [Unreleased]
 ------------
 
+Changed
+^^^^^^^
+* `@lukehsiao`_: Added ``longest_match_only`` parameter to
+  :class:`LambdaFunctionMatcher`, which defaults to False, rather than True.
+  (`#165 <https://github.com/HazyResearch/fonduer/pull/165>`_)
+
 Fixed
 ^^^^^
 * `@lukehsiao`_: Fixes the behavior of the ``get_between_ngrams`` data model
   util. (`#164 <https://github.com/HazyResearch/fonduer/pull/164>`_)
 * `@lukehsiao`_: Batch queries so that PostgreSQL buffers aren't exceeded.
   (`#162 <https://github.com/HazyResearch/fonduer/pull/162>`_)
+
 
 [0.3.2] - 2018-09-20
 --------------------